SAR_WEBSERVICE/requirements.txt
alembic==0.9.6
Babel==2.5.1
bcrypt==3.1.4
blinker==1.4
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
dominate==2.3.1
enum34==1.1.6
Flask==0.12.2
Flask-Babel==0.11.2
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-JWT==0.3.2
Flask-JWT-Extended==3.3.4
Flask-JWT-Simple==0.0.3
Flask-Login==0.4.0
Flask-Mail==0.9.0
Flask-Migrate==2.1.1
Flask-Permissions==0.2.3
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
Flask-User==0.6.19
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.3.2
pycparser==2.18
pycryptodome==3.4.7
PyJWT==1.4.2
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.15
stripe==1.77.1
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
flask-cors==3.0.3
